requirements.txt
altgraph==0.17.4
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.13.1
fsspec==2024.2.0
idna==3.8
Jinja2==3.1.3
keyboard==0.13.5
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.5.0
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai-whisper==20231117
packaging==24.1
pefile==2024.8.26
pillow==10.4.0
plyer==2.1.0
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyMsgBox==1.0.9
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
pywin32==306
pywin32-ctypes==0.2.3
regex==2024.7.24
requests==2.32.3
setuptools==70.0.0
sounddevice==0.5.0
soupsieve==2.6
sympy==1.12
tiktoken==0.7.0
torch==2.4.1+cu124
tqdm==4.66.5
typing_extensions==4.9.0
urllib3==2.2.2
